iox-#25 Remove derivation of PortPool
