+++ iceoryx_posh/include/iceoryx_posh/internal/roudi/memory/port_pool_memory_block.hpp
@@ -14,10 +14,10 @@
 #ifndef IOX_POSH_ROUDI_MEMORY_PORT_POOL_MEMORY_BLOCK_HPP
 #define IOX_POSH_ROUDI_MEMORY_PORT_POOL_MEMORY_BLOCK_HPP
 
+#include "iceoryx_posh/internal/roudi/port_pool_data.hpp"
 #include "iceoryx_posh/mepoo/mepoo_config.hpp"
 #include "iceoryx_posh/roudi/memory/memory_block.hpp"
 #include "iceoryx_utils/cxx/optional.hpp"
-#include "iceoryx_posh/internal/roudi/port_pool_data.hpp"
 
 #include <cstdint>
