+++ iceoryx_posh/include/iceoryx_posh/roudi/port_pool.hpp
@@ -18,11 +18,14 @@
 #include "iceoryx_posh/internal/popo/building_blocks/condition_variable_data.hpp"
 #include "iceoryx_posh/internal/popo/ports/application_port.hpp"
 #include "iceoryx_posh/internal/popo/ports/interface_port.hpp"
+#include "iceoryx_posh/internal/popo/ports/publisher_port_data.hpp"
 #include "iceoryx_posh/internal/popo/ports/publisher_port_roudi.hpp"
+#include "iceoryx_posh/internal/popo/ports/subscriber_port_data.hpp"
 #include "iceoryx_posh/internal/popo/ports/subscriber_port_multi_producer.hpp"
 #include "iceoryx_posh/internal/popo/ports/subscriber_port_single_producer.hpp"
 #include "iceoryx_posh/internal/popo/receiver_port.hpp"
 #include "iceoryx_posh/internal/popo/sender_port.hpp"
+#include "iceoryx_posh/internal/roudi/port_pool_data.hpp"
 #include "iceoryx_posh/internal/runtime/runnable_data.hpp"
 
 namespace iox
@@ -48,7 +51,7 @@ enum class PortPoolError : uint8_t
 class PortPool
 {
   public:
-    PortPool(PortPoolDataBase& portPoolDataBase) noexcept;
+    PortPool(PortPoolData& portPoolData) noexcept;
 
     virtual ~PortPool() noexcept = default;
 
@@ -56,37 +59,36 @@ class PortPool
     /// there could be a member "cxx::vector<popo::SenderPortData* m_senderPorts;" and senderPorts() would just update
     /// this member if the sender ports actually changed
     /// @deprecated #25
-    virtual cxx::vector<SenderPortType::MemberType_t*, MAX_PUBLISHERS> senderPortDataList() noexcept = 0;
+    cxx::vector<SenderPortType::MemberType_t*, MAX_PUBLISHERS> senderPortDataList() noexcept;
     /// @deprecated #25
-    virtual cxx::vector<ReceiverPortType::MemberType_t*, MAX_SUBSCRIBERS> receiverPortDataList() noexcept = 0;
-    virtual cxx::vector<PublisherPortRouDiType::MemberType_t*, MAX_PUBLISHERS> getPublisherPortDataList() noexcept = 0;
-    virtual cxx::vector<SubscriberPortProducerType::MemberType_t*, MAX_SUBSCRIBERS>
-    getSubscriberPortDataList() noexcept = 0;
+    cxx::vector<ReceiverPortType::MemberType_t*, MAX_SUBSCRIBERS> receiverPortDataList() noexcept;
+    cxx::vector<PublisherPortRouDiType::MemberType_t*, MAX_PUBLISHERS> getPublisherPortDataList() noexcept;
+    cxx::vector<SubscriberPortProducerType::MemberType_t*, MAX_SUBSCRIBERS> getSubscriberPortDataList() noexcept;
     cxx::vector<popo::InterfacePortData*, MAX_INTERFACE_NUMBER> getInterfacePortDataList() noexcept;
     cxx::vector<popo::ApplicationPortData*, MAX_PROCESS_NUMBER> getApplicationPortDataList() noexcept;
     cxx::vector<runtime::RunnableData*, MAX_RUNNABLE_NUMBER> getRunnableDataList() noexcept;
 
     /// @deprecated #25
-    virtual cxx::expected<SenderPortType::MemberType_t*, PortPoolError>
+    cxx::expected<SenderPortType::MemberType_t*, PortPoolError>
     addSenderPort(const capro::ServiceDescription& serviceDescription,
                   mepoo::MemoryManager* const memoryManager,
                   const std::string& applicationName,
-                  const mepoo::MemoryInfo& memoryInfo = mepoo::MemoryInfo()) noexcept = 0;
+                  const mepoo::MemoryInfo& memoryInfo = mepoo::MemoryInfo()) noexcept;
 
     /// @deprecated #25
-    virtual cxx::expected<ReceiverPortType::MemberType_t*, PortPoolError>
+    cxx::expected<ReceiverPortType::MemberType_t*, PortPoolError>
     addReceiverPort(const capro::ServiceDescription& serviceDescription,
                     const std::string& applicationName,
-                    const mepoo::MemoryInfo& memoryInfo = mepoo::MemoryInfo()) noexcept = 0;
+                    const mepoo::MemoryInfo& memoryInfo = mepoo::MemoryInfo()) noexcept;
 
-    virtual cxx::expected<PublisherPortRouDiType::MemberType_t*, PortPoolError>
+    cxx::expected<PublisherPortRouDiType::MemberType_t*, PortPoolError>
     addPublisherPort(const capro::ServiceDescription& serviceDescription,
                      const uint64_t& historyCapacity,
                      mepoo::MemoryManager* const memoryManager,
                      const ProcessName_t& applicationName,
-                     const mepoo::MemoryInfo& memoryInfo = mepoo::MemoryInfo()) noexcept = 0;
+                     const mepoo::MemoryInfo& memoryInfo = mepoo::MemoryInfo()) noexcept;
 
-    virtual cxx::expected<SubscriberPortProducerType::MemberType_t*, PortPoolError>
+    cxx::expected<SubscriberPortProducerType::MemberType_t*, PortPoolError>
     addSubscriberPort(const capro::ServiceDescription& serviceDescription,
                       const uint64_t& historyRequest,
                       const ProcessName_t& applicationName,
@@ -104,11 +106,11 @@ class PortPool
     cxx::expected<popo::ConditionVariableData*, PortPoolError> addConditionVariableData() noexcept;
 
     /// @deprecated #25
-    virtual void removeSenderPort(SenderPortType::MemberType_t* const portData) noexcept = 0;
+    void removeSenderPort(SenderPortType::MemberType_t* const portData) noexcept;
     /// @deprecated #25
-    virtual void removeReceiverPort(ReceiverPortType::MemberType_t* const portData) noexcept = 0;
-    virtual void removePublisherPort(PublisherPortRouDiType::MemberType_t* const portData) noexcept = 0;
-    virtual void removeSubscriberPort(SubscriberPortProducerType::MemberType_t* const portData) noexcept = 0;
+    void removeReceiverPort(ReceiverPortType::MemberType_t* const portData) noexcept;
+    void removePublisherPort(PublisherPortRouDiType::MemberType_t* const portData) noexcept;
+    void removeSubscriberPort(SubscriberPortProducerType::MemberType_t* const portData) noexcept;
     void removeInterfacePort(popo::InterfacePortData* const portData) noexcept;
     void removeApplicationPort(popo::ApplicationPortData* const portData) noexcept;
     void removeRunnableData(runtime::RunnableData* const runnableData) noexcept;
@@ -117,7 +119,7 @@ class PortPool
     std::atomic<uint64_t>* serviceRegistryChangeCounter() noexcept;
 
   private:
-    PortPoolDataBase* m_portPoolDataBase;
+    PortPoolData* m_portPoolData;
 };
 
 } // namespace roudi
