+++ iceoryx_posh/source/roudi/port_pool.cpp
@@ -20,32 +20,32 @@ namespace iox
 {
 namespace roudi
 {
-PortPool::PortPool(PortPoolDataBase& portPoolDataBase) noexcept
-    : m_portPoolDataBase(&portPoolDataBase)
+PortPool::PortPool(PortPoolData& portPoolData) noexcept
+    : m_portPoolData(&portPoolData)
 {
 }
 
 cxx::vector<popo::InterfacePortData*, MAX_INTERFACE_NUMBER> PortPool::getInterfacePortDataList() noexcept
 {
-    return m_portPoolDataBase->m_interfacePortMembers.content();
+    return m_portPoolData->m_interfacePortMembers.content();
 }
 
 cxx::vector<popo::ApplicationPortData*, MAX_PROCESS_NUMBER> PortPool::getApplicationPortDataList() noexcept
 {
-    return m_portPoolDataBase->m_applicationPortMembers.content();
+    return m_portPoolData->m_applicationPortMembers.content();
 }
 
 cxx::vector<runtime::RunnableData*, MAX_RUNNABLE_NUMBER> PortPool::getRunnableDataList() noexcept
 {
-    return m_portPoolDataBase->m_runnableMembers.content();
+    return m_portPoolData->m_runnableMembers.content();
 }
 
 cxx::expected<popo::InterfacePortData*, PortPoolError>
 PortPool::addInterfacePort(const std::string& applicationName, const capro::Interfaces interface) noexcept
 {
-    if (m_portPoolDataBase->m_interfacePortMembers.hasFreeSpace())
+    if (m_portPoolData->m_interfacePortMembers.hasFreeSpace())
     {
-        auto interfacePortData = m_portPoolDataBase->m_interfacePortMembers.insert(
+        auto interfacePortData = m_portPoolData->m_interfacePortMembers.insert(
             iox::cxx::string<100>(iox::cxx::TruncateToCapacity, applicationName), interface);
         return cxx::success<popo::InterfacePortData*>(interfacePortData);
     }
@@ -59,9 +59,9 @@ PortPool::addInterfacePort(const std::string& applicationName, const capro::Inte
 cxx::expected<popo::ApplicationPortData*, PortPoolError>
 PortPool::addApplicationPort(const std::string& applicationName) noexcept
 {
-    if (m_portPoolDataBase->m_applicationPortMembers.hasFreeSpace())
+    if (m_portPoolData->m_applicationPortMembers.hasFreeSpace())
     {
-        auto applicationPortData = m_portPoolDataBase->m_applicationPortMembers.insert(
+        auto applicationPortData = m_portPoolData->m_applicationPortMembers.insert(
             iox::cxx::string<100>(iox::cxx::TruncateToCapacity, applicationName));
         return cxx::success<popo::ApplicationPortData*>(applicationPortData);
     }
@@ -75,9 +75,9 @@ PortPool::addApplicationPort(const std::string& applicationName) noexcept
 cxx::expected<runtime::RunnableData*, PortPoolError> PortPool::addRunnableData(
     const ProcessName_t& process, const RunnableName_t& runnable, const uint64_t runnableDeviceIdentifier) noexcept
 {
-    if (m_portPoolDataBase->m_runnableMembers.hasFreeSpace())
+    if (m_portPoolData->m_runnableMembers.hasFreeSpace())
     {
-        auto runnableData = m_portPoolDataBase->m_runnableMembers.insert(process, runnable, runnableDeviceIdentifier);
+        auto runnableData = m_portPoolData->m_runnableMembers.insert(process, runnable, runnableDeviceIdentifier);
         return cxx::success<runtime::RunnableData*>(runnableData);
     }
     else
@@ -87,12 +87,11 @@ cxx::expected<runtime::RunnableData*, PortPoolError> PortPool::addRunnableData(
     }
 }
 
-cxx::expected<popo::ConditionVariableData*, PortPoolError>
-PortPool::addConditionVariableData() noexcept
+cxx::expected<popo::ConditionVariableData*, PortPoolError> PortPool::addConditionVariableData() noexcept
 {
-    if (m_portPoolDataBase->m_conditionVariableMembers.hasFreeSpace())
+    if (m_portPoolData->m_conditionVariableMembers.hasFreeSpace())
     {
-        auto conditionVariableData = m_portPoolDataBase->m_conditionVariableMembers.insert();
+        auto conditionVariableData = m_portPoolData->m_conditionVariableMembers.insert();
         return cxx::success<popo::ConditionVariableData*>(conditionVariableData);
     }
     else
@@ -104,22 +103,145 @@ PortPool::addConditionVariableData() noexcept
 
 void PortPool::removeInterfacePort(popo::InterfacePortData* const portData) noexcept
 {
-    m_portPoolDataBase->m_interfacePortMembers.erase(portData);
+    m_portPoolData->m_interfacePortMembers.erase(portData);
 }
 
 void PortPool::removeApplicationPort(popo::ApplicationPortData* const portData) noexcept
 {
-    m_portPoolDataBase->m_applicationPortMembers.erase(portData);
+    m_portPoolData->m_applicationPortMembers.erase(portData);
 }
 
 void PortPool::removeRunnableData(runtime::RunnableData* const runnableData) noexcept
 {
-    m_portPoolDataBase->m_runnableMembers.erase(runnableData);
+    m_portPoolData->m_runnableMembers.erase(runnableData);
 }
 
 std::atomic<uint64_t>* PortPool::serviceRegistryChangeCounter() noexcept
 {
-    return &m_portPoolDataBase->m_serviceRegistryChangeCounter;
+    return &m_portPoolData->m_serviceRegistryChangeCounter;
+}
+
+/// @deprecated #25
+cxx::vector<SenderPortType::MemberType_t*, MAX_PUBLISHERS> PortPool::senderPortDataList() noexcept
+{
+    return m_portPoolData->m_senderPortMembers.content();
+}
+
+/// @deprecated #25
+cxx::vector<ReceiverPortType::MemberType_t*, MAX_SUBSCRIBERS> PortPool::receiverPortDataList() noexcept
+{
+    return m_portPoolData->m_receiverPortMembers.content();
+}
+
+/// @deprecated #25
+cxx::expected<SenderPortType::MemberType_t*, PortPoolError>
+PortPool::addSenderPort(const capro::ServiceDescription& serviceDescription,
+                        mepoo::MemoryManager* const memoryManager,
+                        const std::string& applicationName,
+                        const mepoo::MemoryInfo& memoryInfo) noexcept
+{
+    if (m_portPoolData->m_senderPortMembers.hasFreeSpace())
+    {
+        auto senderPortData =
+            m_portPoolData->m_senderPortMembers.insert(serviceDescription, memoryManager, applicationName, memoryInfo);
+        return cxx::success<SenderPortType::MemberType_t*>(senderPortData);
+    }
+    else
+    {
+        errorHandler(Error::kPORT_POOL__SENDERLIST_OVERFLOW, nullptr, ErrorLevel::MODERATE);
+        return cxx::error<PortPoolError>(PortPoolError::SENDER_PORT_LIST_FULL);
+    }
+}
+
+/// @deprecated #25
+cxx::expected<ReceiverPortType::MemberType_t*, PortPoolError>
+PortPool::addReceiverPort(const capro::ServiceDescription& serviceDescription,
+                          const std::string& applicationName,
+                          const mepoo::MemoryInfo& memoryInfo) noexcept
+{
+    if (m_portPoolData->m_receiverPortMembers.hasFreeSpace())
+    {
+        auto receiverPortData =
+            m_portPoolData->m_receiverPortMembers.insert(serviceDescription, applicationName, memoryInfo);
+        return cxx::success<ReceiverPortType::MemberType_t*>(receiverPortData);
+    }
+    else
+    {
+        errorHandler(Error::kPORT_POOL__RECEIVERLIST_OVERFLOW, nullptr, ErrorLevel::MODERATE);
+        return cxx::error<PortPoolError>(PortPoolError::RECEIVER_PORT_LIST_FULL);
+    }
+}
+
+/// @deprecated #25
+void PortPool::removeSenderPort(SenderPortType::MemberType_t* const portData) noexcept
+{
+    m_portPoolData->m_senderPortMembers.erase(portData);
+}
+
+/// @deprecated #25
+void PortPool::removeReceiverPort(ReceiverPortType::MemberType_t* const portData) noexcept
+{
+    m_portPoolData->m_receiverPortMembers.erase(portData);
+}
+
+cxx::vector<PublisherPortRouDiType::MemberType_t*, MAX_PUBLISHERS> PortPool::getPublisherPortDataList() noexcept
+{
+    return m_portPoolData->m_publisherPortMembers.content();
+}
+
+cxx::vector<SubscriberPortProducerType::MemberType_t*, MAX_SUBSCRIBERS> PortPool::getSubscriberPortDataList() noexcept
+{
+    return m_portPoolData->m_subscriberPortMembers.content();
+}
+
+cxx::expected<PublisherPortRouDiType::MemberType_t*, PortPoolError>
+PortPool::addPublisherPort(const capro::ServiceDescription& serviceDescription,
+                           const uint64_t& historyCapacity,
+                           mepoo::MemoryManager* const memoryManager,
+                           const ProcessName_t& applicationName,
+                           const mepoo::MemoryInfo& memoryInfo) noexcept
+{
+    if (m_portPoolData->m_publisherPortMembers.hasFreeSpace())
+    {
+        auto publisherPortData = m_portPoolData->m_publisherPortMembers.insert(
+            serviceDescription, applicationName, memoryManager, historyCapacity, memoryInfo);
+        return cxx::success<PublisherPortRouDiType::MemberType_t*>(publisherPortData);
+    }
+    else
+    {
+        errorHandler(Error::kPORT_POOL__PUBLISHERLIST_OVERFLOW, nullptr, ErrorLevel::MODERATE);
+        return cxx::error<PortPoolError>(PortPoolError::PUBLISHER_PORT_LIST_FULL);
+    }
+}
+
+cxx::expected<SubscriberPortProducerType::MemberType_t*, PortPoolError>
+PortPool::addSubscriberPort(const capro::ServiceDescription& serviceDescription,
+                            const uint64_t& historyRequest,
+                            const ProcessName_t& applicationName,
+                            const mepoo::MemoryInfo& memoryInfo) noexcept
+{
+    if (m_portPoolData->m_subscriberPortMembers.hasFreeSpace())
+    {
+        auto subscriberPortData = constructSubscriber<iox::build::CommunicationPolicy>(
+            serviceDescription, historyRequest, applicationName, memoryInfo);
+
+        return cxx::success<SubscriberPortProducerType::MemberType_t*>(subscriberPortData);
+    }
+    else
+    {
+        errorHandler(Error::kPORT_POOL__SUBSCRIBERLIST_OVERFLOW, nullptr, ErrorLevel::MODERATE);
+        return cxx::error<PortPoolError>(PortPoolError::SUBSCRIBER_PORT_LIST_FULL);
+    }
+}
+
+void PortPool::removePublisherPort(PublisherPortRouDiType::MemberType_t* const portData) noexcept
+{
+    m_portPoolData->m_publisherPortMembers.erase(portData);
+}
+
+void PortPool::removeSubscriberPort(SubscriberPortProducerType::MemberType_t* const portData) noexcept
+{
+    m_portPoolData->m_subscriberPortMembers.erase(portData);
 }
 
 } // namespace roudi
